# longaeva/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ContextItemRead": {
        "additionalProperties": false,
        "properties": {
          "observation_id": {
            "format": "uuid",
            "title": "Observation Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "rule_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "rule_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Key"
          }
        },
        "required": [
          "observation_id",
          "reason"
        ],
        "title": "ContextItemRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ParameterEvidence": {
        "additionalProperties": false,
        "properties": {
          "assumption": {
            "default": false,
            "title": "Assumption",
            "type": "boolean"
          },
          "observation_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Observation Ids",
            "type": "array"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          }
        },
        "title": "ParameterEvidence",
        "type": "object"
      },
      "ParameterEvidenceRead": {
        "properties": {
          "evidence": {
            "$ref": "#/components/schemas/ParameterEvidence"
          },
          "excerpts": {
            "items": {
              "$ref": "#/components/schemas/EvidenceExcerpt"
            },
            "title": "Excerpts",
            "type": "array"
          },
          "parameter": {
            "title": "Parameter",
            "type": "string"
          },
          "updates": {
            "items": {
              "$ref": "#/components/schemas/ParameterUpdateDetailRead"
            },
            "title": "Updates",
            "type": "array"
          }
        },
        "required": [
          "parameter",
          "evidence"
        ],
        "title": "ParameterEvidenceRead",
        "type": "object"
      },
...
      "ParameterSetRead": {
        "properties": {
          "assumption_flags": {
            "additionalProperties": true,
            "title": "Assumption Flags",
            "type": "object"
          },
          "company": {
            "title": "Company",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cutoff_ts": {
            "format": "date-time",
            "title": "Cutoff Ts",
            "type": "string"
          },
          "evidence_links": {
            "additionalProperties": true,
            "title": "Evidence Links",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "ranges": {
            "additionalProperties": true,
            "title": "Ranges",
            "type": "object"
          },
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "id",
          "company",
          "cutoff_ts",
          "values",
          "ranges",
          "evidence_links",
          "assumption_flags",
          "content_hash",
          "created_at"
        ],
        "title": "ParameterSetRead",
        "type": "object"
      },
...
      "ParameterUpdateDetailRead": {
        "properties": {
          "after_value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Value"
          },
          "before_value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Value"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "observation_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Observation Ids",
            "type": "array"
          },
          "parameter_set_id": {
            "format": "uuid",
            "title": "Parameter Set Id",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "rule_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "rule_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Key"
          },
          "rule_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Version"
          },
          "size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "target_parameter": {
            "title": "Target Parameter",
            "type": "string"
          }
        },
        "required": [
          "id",
          "parameter_set_id",
          "target_parameter",
          "rationale",
          "observation_ids",
          "created_at"
        ],
        "title": "ParameterUpdateDetailRead",
        "type": "object"
      },
...
      "ProposedUpdateRead": {
        "additionalProperties": false,
        "properties": {
          "after_value": {
            "additionalProperties": true,
            "title": "After Value",
            "type": "object"
          },
          "assumption": {
            "title": "Assumption",
            "type": "boolean"
          },
          "before_value": {
            "additionalProperties": true,
            "title": "Before Value",
            "type": "object"
          },
          "observation_id": {
            "format": "uuid",
            "title": "Observation Id",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "rule_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "rule_key": {
            "title": "Rule Key",
            "type": "string"
          },
          "rule_version": {
            "title": "Rule Version",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "number"
          },
          "target_parameter": {
            "title": "Target Parameter",
            "type": "string"
          }
        },
        "required": [
          "target_parameter",
          "rule_key",
          "rule_version",
          "observation_id",
          "before_value",
          "after_value",
          "size",
          "rationale",
          "assumption"
        ],
        "title": "ProposedUpdateRead",
        "type": "object"
      },
...
      "RuleApplicationRead": {
        "additionalProperties": false,
        "properties": {
          "changed": {
            "title": "Changed",
            "type": "boolean"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/ContextItemRead"
            },
            "title": "Context",
            "type": "array"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "parameter_set_id": {
            "format": "uuid",
            "title": "Parameter Set Id",
            "type": "string"
          },
          "result_parameter_set_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Parameter Set Id"
          },
          "updates": {
            "items": {
              "$ref": "#/components/schemas/ProposedUpdateRead"
            },
            "title": "Updates",
            "type": "array"
          }
        },
        "required": [
          "parameter_set_id",
          "changed",
          "created",
          "updates",
          "context"
        ],
        "title": "RuleApplicationRead",
        "type": "object"
      },
      "RuleApplicationRequest": {
        "additionalProperties": false,
        "properties": {
          "decided_by": {
            "minLength": 1,
            "title": "Decided By",
            "type": "string"
          },
          "families": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Families"
          },
          "observation_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Ids"
          },
          "rationale": {
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "decided_by",
          "rationale"
        ],
        "title": "RuleApplicationRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Standalone Visa business-simulation API (hackathon).",
    "title": "Longaeva",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/parameter-sets": {
      "get": {
        "operationId": "list_parameter_sets_parameter_sets_get",
        "parameters": [
          {
            "in": "query",
            "name": "company",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ParameterSetRead"
                  },
                  "title": "Response List Parameter Sets Parameter Sets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Parameter Sets",
        "tags": [
          "parameter-sets"
        ]
      }
    },
    "/parameter-sets/{parameter_set_id}": {
      "get": {
        "operationId": "get_parameter_set_parameter_sets__parameter_set_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "parameter_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Parameter Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterSetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Parameter Set",
        "tags": [
          "parameter-sets"
        ]
      }
    },
    "/parameter-sets/{parameter_set_id}/apply-rules": {
      "post": {
        "operationId": "apply_parameter_rules_parameter_sets__parameter_set_id__apply_rules_post",
        "parameters": [
          {
            "in": "path",
            "name": "parameter_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Parameter Set Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleApplicationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Parameter Rules",
        "tags": [
          "parameter-sets"
        ]
      }
    },
    "/parameter-sets/{parameter_set_id}/context": {
      "get": {
        "operationId": "list_parameter_context_parameter_sets__parameter_set_id__context_get",
        "parameters": [
          {
            "in": "path",
            "name": "parameter_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Parameter Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContextItemRead"
                  },
                  "title": "Response List Parameter Context Parameter Sets  Parameter Set Id  Context Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Parameter Context",
        "tags": [
          "parameter-sets"
        ]
      }
    },
    "/parameter-sets/{parameter_set_id}/evidence": {
      "get": {
        "operationId": "get_parameter_evidence_parameter_sets__parameter_set_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "parameter_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Parameter Set Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "parameter",
            "required": true,
            "schema": {
              "title": "Parameter",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterEvidenceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Parameter Evidence",
        "tags": [
          "parameter-sets"
        ]
      }
    },
    "/parameter-sets/{parameter_set_id}/lineage": {
      "get": {
        "operationId": "get_parameter_lineage_parameter_sets__parameter_set_id__lineage_get",
        "parameters": [
          {
            "in": "path",
            "name": "parameter_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Parameter Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ParameterSetRead"
                  },
                  "title": "Response Get Parameter Lineage Parameter Sets  Parameter Set Id  Lineage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Parameter Lineage",
        "tags": [
          "parameter-sets"
        ]
      }
    },
    "/parameter-sets/{parameter_set_id}/rule-preview": {
      "post": {
        "operationId": "preview_parameter_rules_parameter_sets__parameter_set_id__rule_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "parameter_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Parameter Set Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleApplicationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Parameter Rules",
        "tags": [
          "parameter-sets"
        ]
      }
    },
    "/parameter-sets/{parameter_set_id}/updates": {
      "get": {
        "operationId": "list_parameter_updates_parameter_sets__parameter_set_id__updates_get",
        "parameters": [
          {
            "in": "path",
            "name": "parameter_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Parameter Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ParameterUpdateDetailRead"
                  },
                  "title": "Response List Parameter Updates Parameter Sets  Parameter Set Id  Updates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Parameter Updates",
        "tags": [
          "parameter-sets"
        ]
      }
    },
...
  }
}